01_software/requirements.txt
pyserial
bottle